backend completed

+++ mycontact-backend/controllers/contactController.js
@@ -3,66 +3,72 @@ const Contact = require("../models/contactModel");
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req, res) => {
-    try {
-      const contacts = await Contact.find();
-      res.status(200).json(contacts);
-    } catch (error) {
-      console.error('Error fetching contacts:', error);
-      res.status(500).json({ message: 'Internal server error' });
-    }
-  });
+  try {
+    const contacts = await Contact.find({ user_id: req.user.id });
+    res.status(200).json(contacts);
+  } catch (error) {
+    console.error('Error fetching contacts:', error);
+    res.status(500).json({ message: 'Internal server error' });
+  }
+});
 
-    //@desc Create New contacts
+//@desc Create New contacts
 //@route POST /api/contacts
-//@access public
-const CreateContact = asyncHandler(async(req,res) => {
-    console.log("The request body is :", req.body);
-    const { name, email, phone} =req.body;
-    if(!name || !email || !phone){
-        res.status(400);
-        throw new Error("All fields are mandatory");
-    }
-    const contact = await Contact.create({
-      name,
-      email,
-      phone,
-    });
-    res.status(201).json(contact);
-    });
-    //@desc Get contacts
+//@access private
+const CreateContact = asyncHandler(async (req, res) => {
+  console.log("The request body is :", req.body);
+  const { name, email, phone } = req.body;
+  if (!name || !email || !phone) {
+    res.status(400);
+    throw new Error("All fields are mandatory");
+  }
+  const contact = await Contact.create({
+    name,
+    email,
+    phone,
+    user_id: req.user.id,
+  });
+  res.status(201).json(contact);
+});
+
+//@desc Get contacts
 //@route GET /api/contacts/:id
-//@access public
-const getContact =asyncHandler(async(req,res) => {
+//@access private
+const getContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
-  if(!contact){
+  if (!contact) {
     res.status(404);
     throw new Error("contact not found");
   }
-    res.status(200).json(contact);
-    });
-    //@desc Update contacts
+  res.status(200).json(contact);
+});
+
+//@desc Update contacts
 //@route PUT /api/contacts
-//@access public
-const UpdateContact = asyncHandler(async(req,res) => {
+//@access private
+const UpdateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
-  if(!contact){
+  if (!contact) {
     res.status(404);
     throw new Error("contact not found");
+  }
+  if (contact.user_id.toString() !== req.user.id.toString()){
+    res.status(403);
+    throw new Error("User dont have permission to update other user contact");
   }
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
-    {new: true}
+    { new: true }
   );
-    res.status(200).json(updatedContact);
-    });
-  
-    //@desc Delete contacts
-//@route DELETE /api/contacts
-//@access public
+  res.status(200).json(updatedContact);
+});
 
+//@desc Delete contacts
+//@route DELETE /api/contacts/:id
+//@access private
 const DeleteContact = asyncHandler(async (req, res) => {
   try {
     const contact = await Contact.findById(req.params.id);
@@ -70,12 +76,24 @@ const DeleteContact = asyncHandler(async (req, res) => {
       res.status(404);
       throw new Error("Contact not found");
     }
-
-    await Contact.deleteOne({ _id: req.params.id });
-    res.status(200).json({ message: "Contact removed" });
+    if (contact.user_id.toString() !== req.user.id.toString()){
+      res.status(403);
+      throw new Error("User dont have permission to update other user contact");
+    }
+    // Delete the contact
+    await Contact.deleteOne( {_id: req.params.id});
+    res.status(200).json({ message: "Contact deleted successfully" });
   } catch (error) {
     console.error('Error deleting contact:', error);
     res.status(500).json({ message: 'Internal server error' });
   }
 });
-    module.exports = {getContacts,CreateContact,getContact,UpdateContact,DeleteContact};
+
+// Exporting all functions
+module.exports = {
+  getContacts,
+  CreateContact,
+  getContact,
+  UpdateContact,
+  DeleteContact
+};
